Extract signup form to own file
